#18-Front-End Interm.Code - Tokens
Tokens are generated by the Scanner of Typee Front-End. They are inputs for the Front-End Parser.
Files describing the tokens, the local tool to be used to automate the generation of the module describing the related nodes of the Intermediaite Code and the very few utilities needed also for this purpose have been added to the repository.